dev(sidecar): stareted using new sidecar
