+++ cmd/insprd/memory/brokers/constants.go
@@ -1,13 +1,13 @@
-package models
+package brokers
 
-import "github.com/inspr/inspr/cmd/insprd/memory/brokers"
+import "github.com/inspr/inspr/pkg/sidecars/models"
 
 // GetSidecarConnectionVars returns port enviroment variable
 // names for each possible broker
-func GetSidecarConnectionVars(broker string) *ConnectionVariables {
+func GetSidecarConnectionVars(broker string) *models.ConnectionVariables {
 	switch broker {
-	case brokers.Kafka:
-		return &ConnectionVariables{
+	case Kafka:
+		return &models.ConnectionVariables{
 			ReadEnvVar:  "INSPR_LBSIDECAR_READ_PORT",
 			WriteEnvVar: "INSPR_SIDECAR_KAFKA_WRITE_PORT",
 		}
