+++ cmd/insprd/memory/fake/factory.go
@@ -1,6 +1,6 @@
 package fake
 
-import "github.com/inspr/inspr/pkg/sidecar_old/models"
+import "github.com/inspr/inspr/pkg/sidecars/models"
 
 // Factory mock of AbsstractBrokersFactory
 type Factory struct {
