+++ cmd/sidecars/kafka/client/reader.go
@@ -6,7 +6,6 @@ import (
 	"github.com/confluentinc/confluent-kafka-go/kafka"
 	globalEnv "github.com/inspr/inspr/pkg/environment"
 	"github.com/inspr/inspr/pkg/ierrors"
-	"github.com/inspr/inspr/pkg/sidecar_old/models"
 	"go.uber.org/zap"
 )
 
@@ -22,33 +21,38 @@ type Consumer interface {
 // Reader reads/commit messages from the channels defined in the env
 type Reader struct {
 	consumers map[string]Consumer
+	kafkaEnv  *Environment
 }
 
 // NewReader return a new Reader
 func NewReader() (*Reader, error) {
 	var reader Reader
+	reader.kafkaEnv = GetEnvironment()
 	channelsList := globalEnv.GetChannelBoundaryList(globalEnv.GetInputChannelsData())
 
 	resolvedChList := globalEnv.GetResolvedBoundaryChannelList(globalEnv.GetInputChannelsData())
 	if len(resolvedChList) == 0 {
-		return nil, ierrors.NewError().Message("KAFKA_INPUT_CHANNELS not specified").InvalidChannel().Build()
+		return nil, ierrors.NewError().Message("INSPR_INPUT_CHANNELS not specified").InvalidChannel().Build()
 	}
 
 	reader.consumers = make(map[string]Consumer)
 
 	for idx, ch := range channelsList {
-		if err := reader.NewSingleChannelConsumer(ch, resolvedChList[idx]); err != nil {
+		if err := reader.newSingleChannelConsumer(ch, resolvedChList[idx]); err != nil {
 			return nil, err
 		}
 	}
 	return &reader, nil
 }
 
-/*
-ReadMessage reads message by message. Returns channel the message belongs to,
-the message and an error if any occurred.
-*/
-func (reader *Reader) ReadMessage(ctx context.Context, channel string) (models.BrokerData, error) {
+// Consumers returns a Reader's consumers
+func (reader *Reader) Consumers() map[string]Consumer {
+	return reader.consumers
+}
+
+// ReadMessage reads message by message. Returns channel the message belongs to,
+// the message and an error if any occurred.
+func (reader *Reader) ReadMessage(ctx context.Context, channel string) ([]byte, error) {
 	resolved, _ := globalEnv.GetResolvedChannel(channel, globalEnv.GetInputChannelsData(), nil)
 
 	logger.Info("trying to read message from topic",
@@ -60,7 +64,7 @@ func (reader *Reader) ReadMessage(ctx context.Context, channel string) (models.B
 	for {
 		select {
 		case <-ctx.Done():
-			return models.BrokerData{}, ctx.Err()
+			return nil, ctx.Err()
 		default:
 			event := consumer.Poll(pollTimeout)
 			switch ev := event.(type) {
@@ -68,19 +72,19 @@ func (reader *Reader) ReadMessage(ctx context.Context, channel string) (models.B
 				topic := *ev.TopicPartition.Topic
 				logger.Info("reading message from topic", zap.String("topic", topic))
 
-				return kafkaTopic(topic).readMessage(ev.Value)
+				return ev.Value, nil
 
 			case kafka.Error:
 				if ev.Code() == kafka.ErrAllBrokersDown {
-					return models.BrokerData{}, ierrors.
+					return nil, ierrors.
 						NewError().
 						InnerError(ev).
 						Message("kafka error = all brokers are down\n%s", ev.Error()).
 						InternalServer().
 						Build()
 				}
-				logger.Error("error in reading kafka message", zap.String("error", ev.Error()))
-				return models.BrokerData{}, ierrors.NewError().
+				logger.Error("error while reading kafka message ...", zap.String("error", ev.Error()))
+				return nil, ierrors.NewError().
 					Message("%v", ev).
 					Build()
 
@@ -105,8 +109,7 @@ func (reader *Reader) Commit(ctx context.Context, channel string) error {
 		if errCommit != nil {
 			return ierrors.
 				NewError().
-				InnerError(errCommit).
-				Message("failed to commit last message").
+				Message("failed to commit last message: %s", errCommit.Error()).
 				InternalServer().
 				Build()
 		}
@@ -126,13 +129,13 @@ func (reader *Reader) Close() error {
 	return nil
 }
 
-//NewSingleChannelConsumer creates a consumer for a single Kafka channel on the reader's consumers map.
-func (reader *Reader) NewSingleChannelConsumer(channel, resolved string) error {
-	kafkaEnv := GetEnvironment()
+//newSingleChannelConsumer creates a consumer for a single Kafka channel on the reader's consumers map.
+func (reader *Reader) newSingleChannelConsumer(channel, resolved string) error {
+
 	newConsumer, errKafkaConsumer := kafka.NewConsumer(&kafka.ConfigMap{
-		"bootstrap.servers":  kafkaEnv.KafkaBootstrapServers,
+		"bootstrap.servers":  reader.kafkaEnv.KafkaBootstrapServers,
 		"group.id":           globalEnv.GetInsprAppID(),
-		"auto.offset.reset":  kafkaEnv.KafkaAutoOffsetReset,
+		"auto.offset.reset":  reader.kafkaEnv.KafkaAutoOffsetReset,
 		"enable.auto.commit": false,
 	})
 	if errKafkaConsumer != nil {
