+++ pkg/environment/mocks.go
@@ -8,7 +8,7 @@ func SetMockEnv() {
 	os.Setenv("INSPR_LBSIDECAR_PORT", "8888")
 	os.Setenv("INSPR_INPUT_CHANNELS", channelsValues)
 	os.Setenv("INSPR_OUTPUT_CHANNELS", channelsValues)
-	os.Setenv("INSPR_SIDECAR_IMAGE", "mock_sidecar_image")
+	os.Setenv("INSPR_LBSIDECAR_IMAGE", "mock_sidecar_image")
 	os.Setenv("INSPR_APP_CTX", "mock.dapp.context")
 	os.Setenv("INSPR_ENV", "mock_env")
 	os.Setenv("INSPR_APP_ID", "testappid1")
@@ -24,7 +24,7 @@ func UnsetMockEnv() {
 	os.Unsetenv("INSPR_INPUT_CHANNELS")
 	os.Unsetenv("INSPR_LBSIDECAR_PORT")
 	os.Unsetenv("INSPR_OUTPUT_CHANNELS")
-	os.Unsetenv("INSPR_SIDECAR_IMAGE")
+	os.Unsetenv("INSPR_LBSIDECAR_IMAGE")
 	os.Unsetenv("INSPR_APP_CTX")
 	os.Unsetenv("INSPR_ENV")
 	os.Unsetenv("INSPR_APP_ID")
